BugFix: Euclidean now works and frontend receives correct TopMathingUsers

+++ backend-assignment-1/src/main/java/lnu/sa224ny/backendassignment1/dtos/SimUserDTO.java
@@ -3,5 +3,5 @@ package lnu.sa224ny.backendassignment1.dtos;
 public class SimUserDTO {
     public String name = null;
     public int id = 0;
-    public float score = 0;
+    public double score = 0;
 }
